Branch Evo Support: IV comparison
isItBetterIV main logic rewritten to support branch evolutions. The
encountered pokemon IVs will be compared with either the top or lower
evos depending on which one it is. This allows the IVs to be compared
with multiple top-level (branched) evos.

Which form it's destined for is determined by evolveMyBest function.